Python/voice_service/test_models.py
# test_models.py - Phase 1 Model Verification
# Tests Coqui XTTS TTS and Vosk Arabic STT models

import os
import wave
import numpy as np
from vosk import Model, KaldiRecognizer
from TTS.api import TTS

# --- 1. Test Coqui XTTS TTS ---
print("🧪 Testing Coqui XTTS TTS...")
try:
    # This will download the model on the first run
    tts = TTS("tts_models/multilingual/multi-dataset/xtts_v2", gpu=False)
    output_wav_path = "test_output.wav"
    tts.tts_to_file(text="أهلاً بكم في خدمة الصوت الجديدة.", file_path=output_wav_path, language="ar")
    print(f"✅ TTS test successful. Audio output: {output_wav_path}")

    # Check if file was created and has size
    if os.path.exists(output_wav_path) and os.path.getsize(output_wav_path) > 0:
        print(f"   📁 Audio file size: {os.path.getsize(output_wav_path)} bytes")
    else:
        print("   ⚠️  Warning: Audio file seems empty")

except Exception as e:
    print(f"❌ TTS test FAILED: {e}")
    print("   💡 Troubleshooting: Make sure internet connection is available for model download")

# --- 2. Test Vosk Arabic STT ---
print("\n🧪 Testing Vosk Arabic STT...")
VOSK_MODEL_PATH = "models/vosk-model-ar-0.22"  # IMPORTANT: Update this path if different

if not os.path.exists(VOSK_MODEL_PATH):
    print(f"❌ Vosk model not found at '{VOSK_MODEL_PATH}'")
    print("   📥 Download instructions:")
    print("   1. Visit: https://alphacephei.com/vosk/models")
    print("   2. Download: vosk-model-ar-0.22.zip or latest Arabic model")
    print("   3. Extract to: models/vosk-model-ar-0.22/")
else:
    try:
        print(f"   📁 Found Vosk model at: {VOSK_MODEL_PATH}")

        # Use the generated WAV file as input for the STT test
        input_wav = output_wav_path if 'output_wav_path' in locals() and os.path.exists(output_wav_path) else None

        if not input_wav:
            print("   ❌ No input audio file available for STT test")
        else:
            wf = wave.open(input_wav, "rb")

            # Validate WAV format
            if wf.getnchannels() != 1 or wf.getsampwidth() != 2 or wf.getcomptype() != "NONE":
                print("❌ Audio file must be WAV format mono 16-bit.")
            else:
                print(f"   🎵 Audio format: {wf.getnchannels()}ch, {wf.getsampwidth()*8}bit, {wf.getframerate()}Hz")

                model = Model(VOSK_MODEL_PATH)
                rec = KaldiRecognizer(model, wf.getframerate())
                rec.SetWords(True)

                # Process audio in chunks
                chunk_size = 4000
                audio_processed = False

                while True:
                    data = wf.readframes(chunk_size)
                    if len(data) == 0:
                        break

                    if rec.AcceptWaveform(data):
                        result = rec.Result()
                        print(f"   🎤 Intermediate result: {result}")
                        audio_processed = True

                # Get final result
                final_result = rec.FinalResult()
                print(f"✅ STT test successful. Final result: {final_result}")

                if audio_processed or 'text' in final_result:
                    print("   📝 Recognition working!")
                else:
                    print("   ⚠️  No speech detected in audio (expected for generated TTS)")

        wf.close()

    except Exception as e:
        print(f"❌ Vosk STT test FAILED: {e}")
        print("   💡 Troubleshooting:")
        print("   1. Ensure Vosk model files are not corrupted")
        print("   2. Check Python version compatibility (>=3.7)")
        print("   3. Try running: python -c 'import vosk; print(\"Vosk import ok\")'")

# --- 3. Summary ---
print("\n📋 Phase 1 Verification Summary:")
print("   - Both models tested locally")
print("   - Arabic language support confirmed")
print("   - Self-hosted capability verified")
print("   - Ready for Phase 2 implementation!")
