Python/voice_service/requirements.txt
# requirements.txt - Python Dependencies for Arabic Healthcare Voice Service
# Generated by uv, managed with uv pip install -r requirements.txt

# Core FastAPI WebSocket server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Voice Processing - Coqui TTS (XTTS v2 model)
coqui-tts>=0.27.0
torch>=2.0.0  # Required for TTS
torchaudio>=2.0.0

# Speech Recognition - Vosk Arabic
vosk==0.3.45

# Audio Processing & Utilities
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1

# Web Framework & CORS
starlette==0.27.0
pydantic==2.5.0
python-multipart==0.0.6

# Optional: Advanced audio processing
soundfile==0.12.1
resampy==0.4.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
rich==13.7.0
loguru==0.7.2

# File system operations
pathlib2==2.3.7

# Configuration
python-dotenv==1.0.0

# WebSocket client for testing (optional)
# websockets-client==1.6.4  # Removed - package doesn't exist in registry

# Health checks
httpx==0.26.0
